added int that was implicit (to eliminate warning)

+++ src/tcl/tclparse.c
@@ -1143,8 +1143,9 @@ Tcl_ParseVar(interp, string, termPtr)
 	    if (TclParseQuotes(interp, string+1, ')', 0, &end, &pv)
 		    != TCL_OK) {
 		char msg[100];
+		/* sde added cast to (int) for precision specifier */
 		sprintf(msg, "\n    (parsing index for array \"%.*s\")",
-			string-name1, name1);
+			(int) (string - name1), name1);
 		Tcl_AddErrorInfo(interp, msg);
 		result = NULL;
 		name2 = pv.buffer;
